tests: Moved code for path generations to object
